+++ examples/python/qb_mpdo_noisy.py
@@ -45,8 +45,8 @@ s.ir_target = circ
 
 # CudaQ has no transpiler, so we need to transpile the circuit to QB's native gate set {rx, ry, cz} first.
 if s.acc[0][0] == "cudaq:qb_mpdo" and s.noise[0][0] == True:
-  # To do this, we simply run the program without executing the circuit, i.e. by setting nosim = True
-  s.nosim = True
+  # To do this, we simply run the program without executing the circuit, i.e. by setting execute_circuit = False
+  s.execute_circuit = False
   s.run()
 
   # Now we can get the transpiled circuit using "out_transpiled_circuit"
@@ -55,7 +55,7 @@ if s.acc[0][0] == "cudaq:qb_mpdo" and s.noise[0][0] == True:
   # The transpiled circuit is in openQasm, so we feed it back into session via "instring"
   s.instring = circ_qasm
 
-  # Execute the transpiled circuit by setting nosim = False
-  s.nosim = False
+  # Execute the transpiled circuit by setting execute_circuit = True
+  s.execute_circuit = True
 s.run()
 print(s.results[0][0])