+++ tests/misc_cpp/transpilationTester.cpp
@@ -25,7 +25,7 @@ h q[1];
   // single CZ.
   my_sim.set_nooptimise(false);
   my_sim.set_noplacement(true);
-  my_sim.set_nosim(true);
+  my_sim.set_execute_circuit(false);
   std::cout << "About to run quantum program..." << std::endl;
   my_sim.run();
 
@@ -56,7 +56,7 @@ cz q[0], q[1];
   // Enable only placement
   my_sim.set_nooptimise(true);
   my_sim.set_noplacement(false);
-  my_sim.set_nosim(true);
+  my_sim.set_execute_circuit(false);
   std::cout << "About to run quantum program..." << std::endl;
   my_sim.run();